fix: Move printer and fixed printer paths
